@@ -1,6 +1,6 @@
 module github.com/amikos-tech/chroma-go/examples/v2/auth
 
-go 1.23.0
+go 1.24
 
 require github.com/amikos-tech/chroma-go v0.2.4
 
@@ -9,7 +9,7 @@ require (
 	github.com/google/uuid v1.6.0 // indirect
 	github.com/oklog/ulid v1.3.1 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
-	github.com/yalue/onnxruntime_go v1.19.0 // indirect
+	github.com/yalue/onnxruntime_go v1.22.0 // indirect
 )
 
 replace github.com/amikos-tech/chroma-go => ../../../
